SwiftCompilerSources: move PhiUpdater.swift from the Optimizer to the SIL module
