+++ SwiftCompilerSources/Sources/Optimizer/PassManager/PassRegistration.swift
@@ -19,7 +19,6 @@ public func initializeSwiftModules() {
   registerAST()
   registerSIL()
   registerSwiftAnalyses()
-  registerUtilities()
   registerSwiftPasses()
   registerOptimizerTests()
 }
@@ -157,8 +156,4 @@ private func registerSwiftPasses() {
 private func registerSwiftAnalyses() {
   AliasAnalysis.register()
   CalleeAnalysis.register()
-}
-
-private func registerUtilities() {
-  registerPhiUpdater()
 }
