+++ SwiftCompilerSources/Sources/Optimizer/TestPasses/CMakeLists.txt
@@ -16,5 +16,6 @@ swift_compiler_sources(Optimizer
   SILPrinter.swift
   RangeDumper.swift
   RunUnitTests.swift
+  UpdateBorrowedFrom.swift
   TestInstructionIteration.swift
 )
