+++ SwiftCompilerSources/Sources/SIL/Utilities/CMakeLists.txt
@@ -10,6 +10,7 @@ swift_compiler_sources(SIL
   AccessUtils.swift
   BorrowUtils.swift
   ForwardingUtils.swift
+  PhiUpdater.swift
   SequenceUtilities.swift
   SmallProjectionPath.swift
   SSAUpdater.swift
